ir_nec: add test for the optional carrier detect feature
The IR-NEC decoder optionally can handle input streams that still have
the carrier signal. Add a test for this configuration, derived from the
test which uses the signal that had the carrier removed in hardware.

The same data is seen, just at different sample numbers. The content was
verified by means of the "cut -d ' ' -f 2-" and "diff" commands.