ir_irmp: introduce IRMP test cases, cover NEC/RC5/RC6/SIRC and others
The upstream IRMP project's decoder core is only slowly moving, we can
consider the IRMP decoder integration to be stable and don't expect the
output to change violently any longer. The issue of requiring a single
core instance remains and affects the GUI, but not the single threaded
test suite.

This set of test cases re-uses the NEC, RC5, RC6, and SIRC dumps which
are covered by individual decoders, too. Ideally detection results would
be identical, but in practise the annotation positions and the level of
details will differ between implementations due to their internal
operation and design choices.

The IRMP test set also covers dumps which are not covered by other IR
decoders. It's interesting to see how not all key repetitions are caught
and how single press might be missed as well. It's valuable to remain
aware during maintenance, and see how occassional failure changes. This
is why an empty output for a non-empty dump is kept here, too.