runtc: add support for input file formats other than native srzip
Introduce the -I command line option for the tstrun(1) utility, accept
input file format names, and options for the input module. Take single
words in multiple -I specs to simplify the internal interface between
the pdtest(1) and the tstrun(1) tools. This follows the existing pattern
for decoder options and channel assignments. The first spec is the file
format name (use 'match' for automatic detection). Subsequent specs are
input module options.

This makes all libsigrok supported input file formats available to the
sigrok-test environment, which formerly was constrained to .sr archives
only. Although the runtc(1) utility resides in the "decoder/" hierarchy,
it can both use VCD and other files to stimulate decoder tests, as well
as cover file formats and their specific features in future tests by
checking that the imported file's content is seen or that their options
take effect.