avclan: Decode new fields and add REPORT_TOC frame
Add a new report frame observed on the bus when a CD changer reports
TOC values from a CD such as total time and number of tracks. This frame
is sent out when a CD is switched.
Also, add clarification on REPORT_LOADER frames where the slot fields
meaning has been further decoded.